Changed the Virtual Machines generation per Service
Virtual Machines now have three maps to store its State Snapshot. The State of a VM consists of the usage of its resources and the instance type.
Implemented a way to keep a single list of Virtual Machines per Service.
Implemented a method to generate end times for the Virtual Machines.